+++ src/scicloj/metamorph/ml/verify.clj
@@ -3,11 +3,11 @@
   (:require
    [clojure.test :refer [is]]
    [scicloj.metamorph.ml :as ml]
-   [scicloj.metamorph.ml.loss :as loss] ;; [tablecloth.api :as tc]
    [tech.v3.dataset :as ds]
    [tech.v3.dataset.column-filters :as cf]
    [tech.v3.dataset.modelling :as ds-mod]
    [tech.v3.datatype.functional :as dfn]
+   [scicloj.metamorph.ml.column-metric :as col-metric]
    [scicloj.metamorph.ml.rdatasets :as rdatasets]))
 
 
@@ -63,12 +63,12 @@
   ([options-map max-avg-loss]
 
    (let [split (ds-mod/train-test-split @regression-iris* options-map)
-         target-colname (first (ds/column-names (cf/target (:test-ds split))))
          train-fn (fn []
                     (let [
                           fitted-model (ml/train (:train-ds split) options-map)
                           predictions (ml/predict (:test-ds split) fitted-model)]
-                      (loss/mae ((:test-ds split) target-colname) (predictions target-colname))))
+                      (col-metric/regression-metric (:test-ds split) predictions :mae)
+                      ))
          avg-mae
          (->>
           (repeatedly 5 train-fn)
@@ -90,20 +90,21 @@
   Returns a clojure.test assertion result.
 
   Used for testing model implementations and ensuring classification models work correctly."
-  ([options-map max-avg-loss]
+  ([options-map min-avg-loss]
 
    (let [split (ds-mod/train-test-split @classification-titanic* options-map)
-         target-colname (first (ds/column-names (cf/target (:test-ds split))))
          train-fn (fn []
                     (let [fitted-model (ml/train (:train-ds split) options-map)
                           predictions (ml/predict (:test-ds split) fitted-model)]
-                      (loss/mae ((:test-ds split) target-colname) (predictions target-colname))))
-         avg-mae
-         (->>
-          (repeatedly 5 train-fn)
-          (dfn/mean))
+                      (col-metric/classification-metric (:test-ds split) predictions :accuracy :macro)
+                      ))
+         avg-accuracy
+         (- 1
+            (->>
+             (repeatedly 5 train-fn)
+             (dfn/mean)))
          ]
-     (is (< avg-mae max-avg-loss))))
+     (is (> avg-accuracy min-avg-loss))))
   ([options-map]
    (basic-classification options-map 0.5)))